add content + styles + upgrade styled-components
