--- components/Content-about.js
@@ -1,18 +1,0 @@
-import React, { Fragment } from 'react'
-import { Content } from '../styles/GlobalStyles'
-
-const About = () => (
-  <Content>
-    Hiding behind the couch until lured out by a feathery toy claws in your leg
-    but shove bum in owner's face like camera lens and ears back wide eyed. Lick
-    left leg for ninety minutes, still dirty lay on arms while you're using the
-    keyboard so twitch tail in permanent irritation scamper russian blue murf
-    pratt ungow ungow, for weigh eight pounds but take up a full-size bed. Chase
-    red laser dot bite nose of your human. Cats secretly make all the worlds
-    muffins meowwww. I could pee on this if i had the energy. You are a captive
-    audience while sitting on the toilet, pet me kitty loves pigs eat the fat
-    cats food.
-  </Content>
-)
-
-export default About
