--- components/Content-contact.js
@@ -1,11 +1,0 @@
-import React, { Fragment } from 'react'
-import { Content } from '../styles/GlobalStyles'
-
-const About = () => (
-  <Fragment>
-    <Content>Feel free to connect with me</Content>
-    <Content>I'll be happy to chat</Content>
-  </Fragment>
-)
-
-export default About
